requirements.txt
scrapy>=2.13.0
sqlalchemy>=2.0.41
beautifulsoup4>=4.13.4
python-dotenv>=1.0.0
python-telegram-bot>=20.7
